+++ src/index.ts
@@ -1,3 +1,6 @@
 export * from "./random";
 export * from "./stack";
 export * from "./getPaths";
+export * from "./loggerObj";
+export * from "./silentLogger";
+export * from "./createTestLogger";