+++ src/silentLogger.ts
@@ -1,0 +1,8 @@
+import type { Logger } from "./loggerObj.js";
+
+export const silentLogger: Logger = {
+  log: () => {},
+  error: () => {},
+  warn: () => {},
+  info: () => {},
+};